solucionando bugs

+++ lestoma.App/lestoma.App/ViewModels/Upas/CreateEditProtocolPopupViewModel.cs
@@ -35,7 +35,7 @@ namespace lestoma.App.ViewModels.Upas
                 if (protocolo != null)
                 {
                     IsEdit = true;
-                    var byteResult = Bytes.Where(x => x == (int)protocolo.PrimerByteTrama).FirstOrDefault();
+                    var byteResult = Bytes.Where(x => x == protocolo.PrimerByteTrama).FirstOrDefault();
                     Protocolo = new ProtocoloModel
                     {
                         Id = protocolo.Id,
